requirements.txt
-i https://pypi.org/simple
asgiref==3.2.3
cffi==1.14.0
cryptography==2.8
django-rest-knox==4.1.0
django==3.0.3
djangorestframework==3.11.0
pycparser==2.19
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
django-heroku==0.3.1
gunicorn==20.0.4
whitenoise==5.0.1